+++ .github/workflows/lint.yml
@@ -25,7 +25,7 @@ jobs:
           pip install flake8 black
 
       - name: Run Flake8
-        run: flake8 --max-line-length 140
+        run: flake8 --max-line-length 150
 
       - name: Run Black
         run: black --check .
